Put MemOutputFile in its own file
