+++ linker/shared/src/main/scala/org/scalajs/linker/LinkerOutput.scala
@@ -70,59 +70,7 @@ final class LinkerOutput private (
 object LinkerOutput extends LinkerOutputPlatformExtensions {
   def apply(jsFile: LinkerOutput.File): LinkerOutput = new LinkerOutput(jsFile)
 
-  def newMemFile(): MemFile = new MemFileImpl()
-
   abstract class File private[linker] () {
     private[linker] def impl: OutputFileImpl
-  }
-
-  sealed trait MemFile extends File {
-    /** Content that has been written to this [[MemFile]].
-     *
-     *  @throws java.lang.IllegalStateException if nothing has been written yet.
-     */
-    def content: Array[Byte]
-  }
-
-  private final class MemFileImpl extends OutputFileImpl with MemFile {
-    @volatile
-    private var _content: Array[Byte] = _
-
-    def content: Array[Byte] = {
-      if (_content == null)
-        throw new IllegalStateException("content hasn't been written yet")
-      _content
-    }
-
-    def newChannel()(implicit ec: ExecutionContext): Future[OutputFileImpl.Channel] =
-      Future.successful(new Channel)
-
-    override def writeFull(buf: ByteBuffer)(implicit ec: ExecutionContext): Future[Unit] = {
-      val c = new Array[Byte](buf.remaining())
-      buf.get(c)
-      _content = c
-      Future.successful(())
-    }
-
-    private class Channel extends OutputFileImpl.Channel {
-      private val out = new ByteArrayOutputStream
-
-      def write(buf: ByteBuffer)(implicit ec: ExecutionContext): Future[Unit] = Future {
-        val promise = Promise[Unit]()
-        if (buf.hasArray()) {
-          out.write(buf.array(), buf.arrayOffset() + buf.position(), buf.remaining())
-          buf.position(buf.limit())
-        } else {
-          val c = new Array[Byte](buf.remaining())
-          buf.get(c)
-          out.write(c)
-        }
-      }
-
-      def close()(implicit ec: ExecutionContext): Future[Unit] = {
-        _content = out.toByteArray
-        Future.successful(())
-      }
-    }
   }
 }
